throws some exceptions in TypeService

+++ src/main/java/com/artplan/pets/service/impl/TypeServiceImpl.java
@@ -9,8 +9,11 @@ import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.artplan.pets.dto.ApiResponse;
 import com.artplan.pets.dto.TypeDto;
 import com.artplan.pets.entity.Type;
+import com.artplan.pets.exception.BadRequestException;
+import com.artplan.pets.exception.ResourceNotFoundException;
 import com.artplan.pets.repository.TypeRepository;
 import com.artplan.pets.service.TypeService;
 
@@ -33,36 +36,43 @@ public class TypeServiceImpl implements TypeService {
 
     @Override
     public List<TypeDto> findAll() {
-        return typeRepository.findAll()
-                .stream()
-                .map(type -> modelMapper.map(type, TypeDto.class))
+        return typeRepository.findAll().stream().map(type -> modelMapper.map(type, TypeDto.class))
                 .collect(Collectors.toList());
     }
 
     @Override
     public TypeDto getById(Long id) {
-        return modelMapper.map(typeRepository.getReferenceById(id), TypeDto.class);
+        Type type = typeRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException(String.format("Type not found with id: '%s'", id)));
+        return modelMapper.map(type, TypeDto.class);
     }
 
     @Override
     public TypeDto add(TypeDto typeDto) {
+        if (Boolean.TRUE.equals(typeRepository.existsByName(typeDto.getName()))) {
+            throw new BadRequestException("Name is already taken");
+        }
         Type type = modelMapper.map(typeDto, Type.class);
-
         type.setId(null);
-
         return modelMapper.map(typeRepository.save(type), TypeDto.class);
     }
 
     @Override
     public TypeDto update(TypeDto typeDto) {
+        if (Boolean.TRUE.equals(typeRepository.existsByName(typeDto.getName()))) {
+            throw new BadRequestException("Name is already taken");
+        }
         Type type = modelMapper.map(typeDto, Type.class);
-
         return modelMapper.map(typeRepository.save(type), TypeDto.class);
     }
 
     @Override
-    public void delete(Long id) {
+    public ApiResponse delete(Long id) {
+        if (Boolean.TRUE.equals(typeRepository.existsById(id))) {
+            throw new ResourceNotFoundException(String.format("Type not found with id: '%s'", id));
+        }
         typeRepository.deleteById(id);
+        return new ApiResponse(true, "You successfully deleted pet");
     }
 
 }
